==> ml-service/src/01_connect_mongo.ipynb <==
from pymongo import MongoClient
import pandas as pd

# Conexión a MongoDB Atlas

# URI de tu cluster real
uri = "mongodb+srv://gabyjargudo82_db_user:BTplZ4YXmOp70kje@horse-trust.4pc35q1.mongodb.net/?appName=horse-trust"

# Conexión al cluster
client = MongoClient(uri)

# Crear/usar base de datos
db = client["horse_trust_db"]

# Crear/usar colección
collection = db["events"]

# Insertar datos no estructurados de ejemplo
sample_data = [
    {"animal": "Caballo", "edad": 5, "color": "Marrón", "observaciones": ["rápido", "juguetón"]},
    {"animal": "Caballo", "edad": 7, "color": "Negro", "observaciones": ["tranquilo", "fuerte"]},
    {"animal": "Caballo", "edad": 3, "color": "Blanco", "observaciones": ["activo", "nervioso"]}
]

# Insertar los datos (solo si la colección está vacía)
if collection.count_documents({}) == 0:
    collection.insert_many(sample_data)
    print("Datos de ejemplo insertados.")
else:
    print("La colección ya tiene datos.")

# Leer los datos de la colección y pasarlos a un DataFrame
data = list(collection.find())
df = pd.DataFrame(data)

# Mostrar el DataFrame
df
