==> ml-service/src/03_EDA.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sd
import os

os.getcwd()

df = pd.read_csv("/workspaces/S02-26-Equipo-30-Web-App-/data/horsetrust_database.csv")

df.head()

df.info()

df.describe()

# Generacion de componentes del indice de confianza Trust Horse

# Completeness score (solo columnas de caballo relevantes) para saber la falta de información

horse_cols = ["h_sex","raza","height_m","weight_kg","length_m","max_speed_kmh",
              "h_temperament","h_category","h_career_races",
              "h_days_since_last_race","h_linaje"]
df["completeness"] = 1 - df[horse_cols].isna().sum(axis=1) / len(horse_cols)

print (df["completeness"])


# Vet Score

df["vet_score"] = np.where(
    df["vet_total_exams"] > 0,
    1 - (df["vet_major_issues"] / df["vet_total_exams"]),
    0.5
)

# Seller score
df["seller_score"] = (
    0.4 * df["s_verified"].astype(int) + 
    0.3 * (1 - np.minimum(df["s_disputes"]/10,1)) +
    0.3 * (1 - df["s_flagged_fraud"].astype(int))
)

# Horse trust score (ponderación final)
df["horse_trust_score"] = 0.4 * df["completeness"] + 0.4 * df["vet_score"] + 0.2 * df["seller_score"]

# Revisar resultados
df[["horse_id","horse_trust_score","completeness","vet_score","seller_score"]].head()

Se ha generado el Horse Trust Score con la pondercion de 3 componentes de informacion del caballo , registro veterinario y 

df.info()

# Nuevo Dataset con el Trust Index

import os
os.getcwd()
os.makedirs("../data", exist_ok=True)
df.to_csv("../data/horsetrust_database_index.csv", index=True)

df.head()
